Split Codex provider tests by concern
